Move SelfRegistering into renderedobject.py
